moved python filed and added doc for preprocessing
