--- sklearn_tda/sklearn_tda/__init__.py
@@ -1,34 +1,0 @@
-name = "sklearn_tda"
-
-from .preprocessing import *
-from .kernel_methods import *
-from .vector_methods import *
-from .metrics import *
-from .clustering import *
-
-__all__ = [
-    "PersistenceImage",
-    "Landscape",
-    "BettiCurve",
-    "Silhouette",
-    "TopologicalVector",
-    "ComplexPolynomial",
-
-    "DiagramSelector",
-    "ProminentPoints",
-    "DiagramPreprocessor",
-    "Padding",
-    "BirthPersistenceTransform",
-
-    "SlicedWassersteinKernel",
-    "PersistenceWeightedGaussianKernel",
-    "PersistenceScaleSpaceKernel",
-    "PersistenceFisherKernel",
-
-    "BottleneckDistance",
-    "SlicedWassersteinDistance",
-    "PersistenceFisherDistance", 
-
-    "MapperComplex",
-    "GraphInducedComplex"
-]
